requirements.txt
pyyaml>=6.0
prometheus-client>=0.20.0
redis>=5.0.0
pytest>=8.0.0
